split out the templates into their own javascript files
