+++ README.md
@@ -5,9 +5,5 @@
 - [ ] Styling Changes
 	- Quotes need vertical space between them when stacked
 - [ ] Text/Wording Changes
-	- get rid of quotes section title?
-	- maybe move quotes title right and de-emphasize?
 	- "Native Fish:" is ugly to me
-- [ ] Code Cleanup
-	- separate javascript by content type (see 3rd to last chapter)
-	- organize
+- [x] Code Cleanup
