+++ styles/quotes.css
@@ -3,3 +3,7 @@
     flex-wrap: wrap;
     justify-content: space-around;
 }
+
+.quote__text {
+	margin: auto;
+}
